+++ applications/fishing-map/src/features/dataviews/dataviews.slice.ts
@@ -1,5 +1,6 @@
 import { createAsyncThunk, createSelector } from '@reduxjs/toolkit'
 import memoize from 'lodash/memoize'
+import { uniqBy } from 'lodash'
 import { Dataview } from '@globalfishingwatch/api-types'
 import GFWAPI from '@globalfishingwatch/api-client'
 import { AsyncReducer, createAsyncSlice } from 'utils/async-slice'
@@ -24,9 +25,12 @@ export const fetchDataviewsByIdsThunk = createAsyncThunk(
     const uniqIds = Array.from(new Set([...ids, ...existingIds]))
     try {
       let dataviews = await GFWAPI.fetch<Dataview[]>(`/v1/dataviews?ids=${uniqIds.join(',')}`)
-      if (process.env.REACT_APP_USE_LOCAL_DATAVIEWS === 'true') {
+      if (
+        process.env.NODE_ENV === 'development' ||
+        process.env.REACT_APP_USE_LOCAL_DATAVIEWS === 'true'
+      ) {
         const mockedDataviews = await import('./dataviews.mock')
-        dataviews = [...dataviews, ...mockedDataviews.default]
+        dataviews = uniqBy([...mockedDataviews.default, ...dataviews], 'id')
       }
       return dataviews
     } catch (e) {
@@ -45,7 +49,7 @@ const { slice: dataviewsSlice, entityAdapter } = createAsyncSlice<ResourcesState
 })
 
 export const {
-  selectAll: selectDataviews,
+  selectAll: selectAllDataviews,
   selectById,
   selectIds,
 } = entityAdapter.getSelectors<RootState>((state) => state.dataviews)