+++ packages/fourwings-aggregate/src/aggregate-tile.ts
@@ -18,6 +18,13 @@ import {
   VALUE_MULTIPLIER,
 } from './constants'
 import { generateUniqueId } from './util'
+import {
+  AggregationOperation,
+  FeatureParams,
+  GeomType,
+  SublayerCombinationMode,
+  TileAggregationParams,
+} from './types'
 
 const getCellCoords = (tileBBox: any, cell: number, numCols: number) => {
   const col = cell % numCols
@@ -33,13 +40,7 @@ const getCellCoords = (tileBBox: any, cell: number, numCols: number) => {
   }
 }
 
-const getPointFeature = (
-  tileBBox: any,
-  cell: number,
-  numCols: number,
-  numRows: number,
-  addMeta: boolean
-): any => {
+const getPointFeature = ({ tileBBox, cell, numCols, numRows, addMeta }: FeatureParams): any => {
   const [minX, minY] = tileBBox
   const { col, row, width, height } = getCellCoords(tileBBox, cell, numCols)
 
@@ -63,13 +64,7 @@ const getPointFeature = (
   }
 }
 
-const getRectangleFeature = (
-  tileBBox: any,
-  cell: number,
-  numCols: number,
-  numRows: number,
-  addMeta: boolean
-): any => {
+const getRectangleFeature = ({ tileBBox, cell, numCols, numRows, addMeta }: FeatureParams): any => {
   const [minX, minY] = tileBBox
   const { col, row, width, height } = getCellCoords(tileBBox, cell, numCols)
 
@@ -103,29 +98,13 @@ const getRectangleFeature = (
   }
 }
 
-const getFeature = ({
-  geomType,
-  tileBBox,
-  cell,
-  numCols,
-  numRows,
-  id,
-  addMeta,
-}: {
-  geomType: string
-  tileBBox: any
-  cell: number
-  numCols: number
-  numRows: number
-  id: string
-  addMeta: boolean
-}) => {
+const getFeature = (featureParams: FeatureParams) => {
   const feature =
-    geomType === 'point'
-      ? getPointFeature(tileBBox, cell, numCols, numRows, addMeta)
-      : getRectangleFeature(tileBBox, cell, numCols, numRows, addMeta)
+    featureParams.geomType === GeomType.point
+      ? getPointFeature(featureParams)
+      : getRectangleFeature(featureParams)
 
-  feature.id = id
+  feature.id = featureParams.id
 
   return feature
 }
@@ -176,7 +155,7 @@ const getBucketIndex = (breaks: number[], value: number) => {
   return currentBucketIndex
 }
 
-const getAddValue = (realValuesSum: number, breaks: number[][]) => {
+const getValue = (realValuesSum: number, breaks?: number[][]) => {
   if (realValuesSum === 0) return undefined
   return breaks ? getBucketIndex(breaks[0], realValuesSum) : realValuesSum
 }
@@ -184,7 +163,7 @@ const getAddValue = (realValuesSum: number, breaks: number[][]) => {
 const getCompareValue = (
   datasetsHighestRealValue: number,
   datasetsHighestRealValueIndex: number,
-  breaks: number[][]
+  breaks?: number[][]
 ) => {
   if (datasetsHighestRealValue === 0) return undefined
   if (breaks) {
@@ -199,7 +178,7 @@ const getCompareValue = (
   }
 }
 
-const getBivariateValue = (realValues: number[], breaks: number[][]) => {
+const getBivariateValue = (realValues: number[], breaks?: number[][]) => {
   if (realValues[0] === 0 && realValues[1] === 0) return undefined
   if (breaks) {
     //  y: datasetB
@@ -231,8 +210,8 @@ const getBivariateValue = (realValues: number[], breaks: number[][]) => {
   }
 }
 
-const getLiteralValues = (realValues: number[], numDatasets: number) => {
-  if (numDatasets === 1) return realValues
+const getLiteralValues = (realValues: number[], sublayerCount: number) => {
+  if (sublayerCount === 1) return realValues
   return `[${realValues.join(',')}]`
 }
 
@@ -241,45 +220,53 @@ const getCumulativeValue = (realValuesSum: number, cumulativeValuesPaddedStrings
   return cumulativeValuesPaddedStrings.join('')
 }
 
-const aggregate = (intArray: number[], options: any) => {
+const aggregate = (intArray: number[], options: TileAggregationParams) => {
   const {
     quantizeOffset = 0,
     tileBBox,
-    delta = 30,
-    geomType = 'rectangle',
-    singleFrame,
-    interactive,
-    breaks,
     x,
     y,
     z,
-    numDatasets,
-    combinationMode,
-    visible,
+    delta = 30,
+    geomType = GeomType.rectangle,
+    singleFrame,
+    interactive,
+    sublayerBreaks,
+    sublayerCount,
+    sublayerCombinationMode,
+    sublayerVisibility,
+    aggregationOperation,
   } = options
-
+  if (sublayerCombinationMode === SublayerCombinationMode.None && sublayerCount > 1) {
+    throw new Error('Multiple sublayers but no proper combination mode set')
+  }
   if (
-    breaks &&
-    breaks.length !== numDatasets &&
-    (combinationMode === 'compare' || combinationMode === 'bivariate')
+    sublayerBreaks &&
+    sublayerBreaks.length !== sublayerCount &&
+    (sublayerCombinationMode === SublayerCombinationMode.Max ||
+      sublayerCombinationMode === SublayerCombinationMode.Bivariate)
   ) {
     throw new Error(
       'must provide as many breaks arrays as number of datasets when using compare and bivariate modes'
     )
   }
-  if (breaks && breaks.length !== 1 && combinationMode === 'add') {
+  if (
+    sublayerBreaks &&
+    sublayerBreaks.length !== 1 &&
+    sublayerCombinationMode === SublayerCombinationMode.Add
+  ) {
     throw new Error('add combinationMode requires one and only one breaks array')
   }
-  if (combinationMode === 'bivariate') {
-    if (numDatasets !== 2)
+  if (sublayerCombinationMode === SublayerCombinationMode.Bivariate) {
+    if (sublayerCount !== 2)
       throw new Error('bivariate combinationMode requires exactly two datasets')
-    if (breaks) {
-      if (breaks.length !== 2)
+    if (sublayerBreaks) {
+      if (sublayerBreaks.length !== 2)
         throw new Error('bivariate combinationMode requires exactly two breaks array')
-      if (breaks[0].length !== breaks[1].length)
+      if (sublayerBreaks[0].length !== sublayerBreaks[1].length)
         throw new Error('bivariate breaks arrays must have the same length')
       // TODO This might change if we want bivariate with more or less than 16 classes
-      if (breaks[0].length !== 4 || breaks[1].length !== 4)
+      if (sublayerBreaks[0].length !== 4 || sublayerBreaks[1].length !== 4)
         throw new Error('each bivariate breaks array require exactly 4 values')
     }
   }
@@ -287,8 +274,9 @@ const aggregate = (intArray: number[], options: any) => {
   const features = []
   const featuresInteractive = []
 
-  let aggregating = Array(numDatasets).fill([])
-  let currentAggregatedValues = Array(numDatasets).fill(0)
+  let aggregating = Array(sublayerCount).fill([])
+  let currentAggregatedValues = Array(sublayerCount).fill(0)
+  let currentAggregatedValuesLength = 0
 
   let currentFeature
   let currentFeatureInteractive
@@ -316,6 +304,10 @@ const aggregate = (intArray: number[], options: any) => {
   let endIndex = 0
   let indexInCell = 0
 
+  // We need to pad with n values (n === delta) to generate "overflow" frames
+  // in the case of a sum, add zeroes which will get added to the running sunm with no effect
+  // in the case of avg, us NaN as a flag to not take the value into account
+  const padValue = aggregationOperation === AggregationOperation.Avg ? NaN : 0
   for (let i = FEATURE_CELLS_START_INDEX; i < intArray.length; i++) {
     const value = intArray[i]
     if (indexInCell === CELL_NUM_INDEX) {
@@ -325,13 +317,15 @@ const aggregate = (intArray: number[], options: any) => {
       startFrame = value
     } else if (indexInCell === CELL_END_INDEX) {
       endFrame = value
-      endIndex = startIndex + CELL_VALUES_START_INDEX + (endFrame - startFrame + 1) * numDatasets
+      endIndex = startIndex + CELL_VALUES_START_INDEX + (endFrame - startFrame + 1) * sublayerCount
     }
     indexInCell++
     if (i === endIndex - 1) {
       indexInCell = 0
       const original = intArray.slice(startIndex, endIndex)
-      const padded = new Array(delta * numDatasets).fill(0)
+      const padded = new Array(delta * sublayerCount).fill(padValue)
+
+      // TODO Are we sure we want to use FEATURE_CELLS_START_INDEX, not CELL_START_INDEX??
       original[FEATURE_CELLS_START_INDEX] = endFrame + delta
       const merged = original.concat(padded)
       featureIntArrays.push(merged)
@@ -339,7 +333,6 @@ const aggregate = (intArray: number[], options: any) => {
   }
 
   // const t = performance.now()
-  // console.log(x, y, z, intArray)
 
   if (singleFrame) {
     for (let i = 2; i < intArray.length; i++) {
@@ -351,12 +344,12 @@ const aggregate = (intArray: number[], options: any) => {
         const featureParams = {
           geomType,
           tileBBox,
-          cell: currentFeatureCell,
+          cell: currentFeatureCell as number,
           numCols,
           numRows,
           id: uniqueId,
         }
-        currentFeature = getFeature(featureParams as any)
+        currentFeature = getFeature(featureParams)
         currentFeature.properties.value = value / VALUE_MULTIPLIER
         features.push(currentFeature)
       }
@@ -378,9 +371,9 @@ const aggregate = (intArray: number[], options: any) => {
         id: uniqueId,
         addMeta: true,
       }
-      currentFeature = getFeature(featureParams as any)
+      currentFeature = getFeature(featureParams)
       if (interactive) {
-        currentFeatureInteractive = getFeature({ ...featureParams, addMeta: true } as any)
+        currentFeatureInteractive = getFeature({ ...featureParams, addMeta: true })
       }
 
       for (let i = CELL_VALUES_START_INDEX; i < featureIntArray.length; i++) {
@@ -391,7 +384,7 @@ const aggregate = (intArray: number[], options: any) => {
 
         // gets index of dataset, knowing that after headers values go
         // dataset1, dataset2, dataset1, dataset2, ...
-        const datasetIndex = featureBufferValuesPos % numDatasets
+        const datasetIndex = featureBufferValuesPos % sublayerCount
 
         // collect value for this dataset
         aggregating[datasetIndex].push(value)
@@ -399,69 +392,89 @@ const aggregate = (intArray: number[], options: any) => {
         let tailValue = 0
         if (tail > currentFeatureMinTimestamp) {
           tailValue = aggregating[datasetIndex].shift()
+        } else {
+          currentAggregatedValuesLength++
         }
 
         // collect "working" value, ie value at head by substracting tail value
         let realValueAtFrameForDataset = 0
-        if (visible[datasetIndex]) {
-          realValueAtFrameForDataset = currentAggregatedValues[datasetIndex] + value - tailValue
+        let realValueAtFrameForDatasetWorkingValue = 0
+        if (sublayerVisibility[datasetIndex]) {
+          if (aggregationOperation === AggregationOperation.Avg) {
+            // if isNan, value is just for padding - stop incrementing running sum (just remove tail)
+            // and take into account one less frame to compute teh avg
+            realValueAtFrameForDatasetWorkingValue = isNaN(value)
+              ? currentAggregatedValues[datasetIndex] - tailValue
+              : currentAggregatedValues[datasetIndex] + value - tailValue
+            if (isNaN(value)) currentAggregatedValuesLength--
+            realValueAtFrameForDataset =
+              realValueAtFrameForDatasetWorkingValue / currentAggregatedValuesLength
+          } else {
+            realValueAtFrameForDataset = realValueAtFrameForDatasetWorkingValue =
+              currentAggregatedValues[datasetIndex] + value - tailValue
+          }
         }
-        currentAggregatedValues[datasetIndex] = realValueAtFrameForDataset
+        currentAggregatedValues[datasetIndex] = realValueAtFrameForDatasetWorkingValue
 
         // Compute mode-specific values
-        if (combinationMode === 'compare') {
+        if (sublayerCombinationMode === SublayerCombinationMode.Max) {
           if (realValueAtFrameForDataset > datasetsHighestRealValue) {
             datasetsHighestRealValue = realValueAtFrameForDataset
             datasetsHighestRealValueIndex = datasetIndex
           }
         }
-        if (combinationMode === 'add' || combinationMode === 'cumulative') {
+        if (
+          sublayerCombinationMode === SublayerCombinationMode.Add ||
+          sublayerCombinationMode === SublayerCombinationMode.Cumulative
+        ) {
           realValuesSum += realValueAtFrameForDataset
         }
-        if (combinationMode === 'cumulative') {
+        if (sublayerCombinationMode === SublayerCombinationMode.Cumulative) {
           const cumulativeValuePaddedString = Math.round(realValuesSum).toString().padStart(6, '0')
           cumulativeValuesPaddedStrings.push(cumulativeValuePaddedString)
         }
-        if (combinationMode === 'literal') {
+        if (sublayerCombinationMode === SublayerCombinationMode.Literal) {
           // literalValuesStr += Math.floor(realValueAtFrameForDataset * 100) / 100
           // Just rounding is faster - revise if decimals are needed
           // Use ceil to avoid values being 'mute' when very close to zero
           // Update: use .round to avoid discrepancies betwen interaction and total ammount
           literalValuesStr += Math.round(realValueAtFrameForDataset)
-          if (datasetIndex < numDatasets - 1) {
+          if (datasetIndex < sublayerCount - 1) {
             literalValuesStr += ','
           }
         }
 
         const quantizedTail = tail - quantizeOffset
 
-        if (quantizedTail >= 0 && datasetIndex === numDatasets - 1) {
+        if (quantizedTail >= 0 && datasetIndex === sublayerCount - 1) {
           let finalValue
 
-          if (combinationMode === 'literal') {
+          if (sublayerCombinationMode === SublayerCombinationMode.Literal) {
             literalValuesStr += ']'
           }
-          // TODO add 'single' mode
-          if (combinationMode === 'compare') {
+
+          if (sublayerCombinationMode === SublayerCombinationMode.None) {
+            finalValue = getValue(realValueAtFrameForDataset, sublayerBreaks)
+          } else if (sublayerCombinationMode === SublayerCombinationMode.Max) {
             finalValue = getCompareValue(
               datasetsHighestRealValue,
               datasetsHighestRealValueIndex as number,
-              breaks
+              sublayerBreaks
             )
-          } else if (combinationMode === 'add') {
-            finalValue = getAddValue(realValuesSum, breaks)
-          } else if (combinationMode === 'bivariate') {
-            finalValue = getBivariateValue(currentAggregatedValues, breaks)
-          } else if (combinationMode === 'literal') {
+          } else if (sublayerCombinationMode === SublayerCombinationMode.Add) {
+            finalValue = getValue(realValuesSum, sublayerBreaks)
+          } else if (sublayerCombinationMode === SublayerCombinationMode.Bivariate) {
+            finalValue = getBivariateValue(currentAggregatedValues, sublayerBreaks)
+          } else if (sublayerCombinationMode === SublayerCombinationMode.Literal) {
             finalValue = literalValuesStr
-          } else if (combinationMode === 'cumulative') {
+          } else if (sublayerCombinationMode === SublayerCombinationMode.Cumulative) {
             finalValue = getCumulativeValue(realValuesSum, cumulativeValuesPaddedStrings)
           }
           // console.log(quantizedTail, finalValue, currentFeature)
           writeValueToFeature(quantizedTail, finalValue as string | number, currentFeature)
         }
 
-        if (datasetIndex === numDatasets - 1) {
+        if (datasetIndex === sublayerCount - 1) {
           // When all dataset values have been collected for this frame, we can move to next frame
           head++
 
@@ -487,8 +500,8 @@ const aggregate = (intArray: number[], options: any) => {
       realValuesSum = 0
       cumulativeValuesPaddedStrings = []
 
-      aggregating = Array(numDatasets).fill([])
-      currentAggregatedValues = Array(numDatasets).fill(0)
+      aggregating = Array(sublayerCount).fill([])
+      currentAggregatedValues = Array(sublayerCount).fill(0)
 
       continue
     }