+++ packages/fourwings-aggregate/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/fourwings-aggregate
 
+## 3.1.0
+
+### Minor Changes
+
+- export consts
+
+## 3.0.0
+
+### Major Changes
+
+- 358004cb: Use client-side aggregation with env layers
+
 ## 1.2.1
 
 ### Patch Changes