+++ packages/ocean-areas/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 0.0.1
 
+## 0.2.2
+
+### Patch Changes
+
+- support bahasa indonesia language
+
 ## 0.2.1
 
 ### Patch Changes